% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_cs.R
\name{cs-methods}
\alias{cs-methods}
\title{Compressed Sensing (CS) unfolding}
\description{
R port of \code{bssunfold/src/bssunfold/core/unfold_cs.py}.
Implements a neutron spectrum unfolding method based on compressive
sensing. The spectrum \eqn{x} is represented sparsely in a dictionary
\eqn{D} as \eqn{x = D \alpha}. The measurement equation \eqn{b = A x}
becomes \eqn{b = (A D) \alpha}, which is solved for the sparse \eqn{\alpha}
using SL0. Finally the spectrum is reconstructed as \eqn{x = D \alpha}.
}
\details{
This package provides the OMP and SL0 sub-solvers and the main CS wrapper.
K-SVD dictionary learning is also included but not used by the main wrapper
(which uses the trivial identity dictionary by default).
}
