% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_ensemble.R
\docType{data}
\name{.bss_cascade_dispatch}
\alias{.bss_cascade_dispatch}
\title{METHOD_DISPATCH from unfold_cascade.py (short name -> Detector.unfold_*).}
\format{
An object of class \code{character} of length 23.
}
\usage{
.bss_cascade_dispatch
}
\description{
METHOD_DISPATCH from unfold_cascade.py (short name -> Detector.unfold_*).
}
\keyword{internal}
