% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_ensemble.R
\docType{data}
\name{.bss_composite_base_weights}
\alias{.bss_composite_base_weights}
\title{DEFAULT_ENSEMBLE_WEIGHTS from unfold_composite.py.}
\format{
An object of class \code{numeric} of length 22.
}
\usage{
.bss_composite_base_weights
}
\description{
DEFAULT_ENSEMBLE_WEIGHTS from unfold_composite.py.
}
\keyword{datasets}
