% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_ensemble.R
\docType{data}
\name{.bss_composite_dispatch}
\alias{.bss_composite_dispatch}
\title{METHOD_DISPATCH from unfold_composite.py.}
\format{
An object of class \code{character} of length 22.
}
\usage{
.bss_composite_dispatch
}
\description{
METHOD_DISPATCH from unfold_composite.py.
}
\keyword{datasets}
\keyword{internal}
