% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_ensemble.R
\name{.bss_select_next_method}
\alias{.bss_select_next_method}
\title{select_next_method() from unfold_cascade.py:245.}
\usage{
.bss_select_next_method(current_metrics, available_methods, stage_number)
}
\description{
select_next_method() from unfold_cascade.py:245.
}
\keyword{internal}
