% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_qp_commercial.R
\name{.commercial_cosine_similarity_selection}
\alias{.commercial_cosine_similarity_selection}
\title{Cosine-similarity alpha selection (\code{cosine_similarity_selection})}
\usage{
.commercial_cosine_similarity_selection(
  A,
  b,
  initial_spectrum,
  n_alphas = 100L,
  alpha_range = c(-9, 2),
  norm = 2L
)
}
\description{
Cosine-similarity alpha selection (\code{cosine_similarity_selection})
}
\keyword{internal}
