% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_qp_commercial.R
\name{.commercial_estimate_noise_variance}
\alias{.commercial_estimate_noise_variance}
\title{Population noise variance of the least-squares residual}
\usage{
.commercial_estimate_noise_variance(A, b)
}
\description{
Population noise variance of the least-squares residual
}
\keyword{internal}
