% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_qp_commercial.R
\name{.commercial_select_regularization_parameter}
\alias{.commercial_select_regularization_parameter}
\title{Dispatch for the automatic selection methods}
\usage{
.commercial_select_regularization_parameter(
  A,
  b,
  method,
  noise_var = NULL,
  initial_spectrum = NULL
)
}
\description{
Dispatch for the automatic selection methods
}
\keyword{internal}
