% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.to_df}
\alias{.to_df}
\title{Extract a tidy data frame from an unfold result}
\usage{
.to_df(result)
}
\arguments{
\item{result}{A result list with \code{energy} and \code{spectrum} elements.}
}
\description{
Extract a tidy data frame from an unfold result
}
\keyword{internal}
