% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_mlem_bs.R
\name{ks_statistic}
\alias{ks_statistic}
\title{K_S goodness-of-fit statistic}
\usage{
ks_statistic(measurements, estimate)
}
\arguments{
\item{measurements}{Numeric vector.}

\item{estimate}{Numeric vector (model prediction).}
}
\value{
Numeric scalar.
}
\description{
\eqn{K_S = | sum_i (n_i - model_i)^2 / sum_i model_i - 1 |}
}
\examples{
ks_statistic(c(1, 2, 3), c(1.1, 1.9, 3.1))
}
