% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_optimization.R
\name{optimization-methods}
\alias{optimization-methods}
\title{LMfit / QP / CVXPY-equivalent unfolding methods}
\description{
R ports of \code{bssunfold/src/bssunfold/core/unfold_lmfit.py},
\code{unfold_qpsolvers.py}, \code{unfold_cvxpy.py}.
}
