% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot measurement residuals of an unfold result}
\usage{
plot_residuals(result, detector_names = NULL)
}
\arguments{
\item{result}{A result list (\code{effective_readings},
\code{residual}).}

\item{detector_names}{Optional names of the detectors (used for
x-labelling).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot measurement residuals of an unfold result
}
