% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_binned.R
\name{save_bin_lookup}
\alias{save_bin_lookup}
\alias{load_bin_lookup}
\title{Save / load bin lookup}
\usage{
save_bin_lookup(lookup, path)

load_bin_lookup(path)
}
\arguments{
\item{lookup}{Integer vector.}

\item{path}{File path.}
}
\value{
Invisible \code{lookup} (for save) or the loaded integer vector.
}
\description{
Save / load bin lookup
}
